+++ app/src/main/java/nextstep/joanne/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package nextstep.joanne.domain;
+package nextstep.joanne.http;
 
 import java.util.Objects;
